requirements.txt
absl-py==0.15.0
ase==3.21.1
astunparse==1.6.3
bayesian-optimization==1.3.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
cycler==0.10.0
decorator==4.4.2
dgl-cu110 @ file:///gpfs1/home/jzhang988/software/dgl_cu110-0.7.1-cp37-cp37m-manylinux1_x86_64.whl
dscribe==1.2.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.10.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.39.1
Markdown==3.3.6
matplotlib==3.4.2
monty==2021.12.1
mpmath==1.2.1
networkx==2.6.3
numba==0.56.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
palettable==3.3.0
pandas==1.3.5
Pillow==8.2.0
plotly==5.5.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
PyDispatcher==2.0.5
pymatgen==2020.11.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.6.3
shiboken2==5.15.2
six==1.15.0
sklearn==0.0
sparse==0.13.0
spglib==1.16.3
sympy==1.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.1
typing-extensions==3.7.4.3
uncertainties==3.1.6
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
